requirements.txt
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0
SQLAlchemy>=2.0.0
pyodbc>=4.0.39
